Initial Commit - Prototype / Proof of Concept
Character is animated using a custom script to load the torso/body
sprites and cycle between the animation. The player controller is
driving the animator class, but some refactoring with design patterns is
needed to clean up the animator.. Additionally, the caching mechanism
will need to be more efficient to handle many characters on the screen.